+++ core/src/main/java/com/adobe/aem/commons/assetshare/content/renditions/impl/dispatchers/ExternalRedirectRenditionDispatcherImpl.java
@@ -90,7 +90,11 @@ public class ExternalRedirectRenditionDispatcherImpl extends AbstractRenditionDi
 
     @Override
     public Set<String> getRenditionNames() {
-        return mappings.keySet();
+        if (mappings == null) {
+            return Collections.EMPTY_SET;
+        } else {
+            return mappings.keySet();
+        }
     }
 
     @Override